decouple cgo code from getaddress
